# Conflicts:
#	ShellCrash.tar.gz
#	bin/version
#	scripts/menus/5_task.sh
#	version

+++ scripts/starts/snapshot_init.sh
@@ -71,11 +71,11 @@ auto_start(){
 		set_profile '/etc/profile' 
 	}
 	#启动服务
-	if [ ! -f "$CRASHDIR"/.dis_startup ]; then
+	if [ ! -f "$CRASHDIR"/.dis_startup ] && [ ! -f "$CRASHDIR"/.start_error ]; then
 		#AX6S/AX6000修复tun功能
 		[ -s "$CRASHDIR"/tools/tun.ko ] && tunfix
 		#小米7000/小米万兆修复tproxy
-		[ -f /etc/init.d/qca-nss-ecm ] && [ "$redir_mod" = 'Tproxy' ] && tproxyfix
+		[ -f /etc/init.d/qca-nss-ecm ] && echo "$redir_mod" | grep -q 'Tproxy' && tproxyfix
 		#自动覆盖根证书文件
 		[ -s "$CRASHDIR"/tools/ca-certificates.crt ] && cp -f "$CRASHDIR"/tools/ca-certificates.crt /etc/ssl/certs/ca-certificates.crt
 		#启动服务
